salamgcc/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ad-extractor"
version = "1.0.0"
description = "Multi-platform classified ads scraper for car listings (Blocket.se)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "SalaamGCC"}]
keywords = ["scraper", "blocket", "cars", "classifieds", "mongodb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "setuptools>=65.0,<82",  # pkg_resources removed in 82; selenium-wire needs it
    "selenium>=4.16.0",
    "blinker<1.8.0",  # selenium-wire requires blinker._saferef (removed in 1.8+)
    "selenium-wire>=5.1.0",
    "webdriver-manager>=4.0.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "deep-translator>=1.11.0",
    "requests>=2.31.0",
    "fake-useragent>=1.4.0",
    "pymongo>=4.6.0",
    "dnspython>=2.4.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pillow>=10.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
ad-extractor = "ad_extractor.cli:main"

[project.urls]
Repository = "https://github.com/SalaamGCC/ad-extractor-python"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

